requirements.txt
Flask>=2.0
Flask-SocketIO>=5.0
psutil>=5.8
eventlet>=0.30
pywebview>=4.0